adding data in read_batch for faenet
