+++ .gitignore
@@ -112,6 +112,3 @@ Local
 
 # VS Code
 .vscode/
-
-
-
