+++ matsciml/models/pyg/faenet/faenet.py
@@ -5,360 +5,16 @@ Simple, scalable and expressive model for property prediction on 3D atomic syste
 from __future__ import annotations
 
 from copy import deepcopy
-from typing import Dict, Optional, Union
 
 import torch
 from torch import nn
-from torch.nn import Embedding, Linear
-from torch_geometric.nn import LayerNorm, MessagePassing
-from torch_geometric.nn.norm import GraphNorm
-from torch_geometric.nn.pool import global_add_pool
-from torch_geometric.typing import Size
+from torch.nn import Linear
 
-from matsciml.common.types import AbstractGraph
+from matsciml.common.types import AbstractGraph, BatchDict, DataDict
+from matsciml.common.utils import radius_graph_pbc
 from matsciml.models.base import AbstractPyGModel
 from matsciml.models.pyg.faenet.helper import *
 from matsciml.models.pyg.faenet.layers import *
-
-
-class EmbeddingBlock(nn.Module):
-    """Initialise atom and edge representations."""
-
-    def __init__(
-        self,
-        num_gaussians,
-        num_filters,
-        hidden_channels,
-        tag_hidden_channels,
-        pg_hidden_channels,
-        phys_hidden_channels,
-        phys_embeds,
-        act,
-        second_layer_MLP,
-    ):
-        super().__init__()
-        self.act = act
-        self.use_tag = tag_hidden_channels > 0
-        self.use_pg = pg_hidden_channels > 0
-        self.use_mlp_phys = phys_hidden_channels > 0 and phys_embeds
-        self.second_layer_MLP = second_layer_MLP
-
-        # --- Node embedding ---
-
-        # Phys embeddings
-        self.phys_emb = PhysEmbedding(
-            props=phys_embeds,
-            props_grad=phys_hidden_channels > 0,
-            pg=self.use_pg,
-        )
-        # With MLP
-        if self.use_mlp_phys:
-            self.phys_lin = Linear(self.phys_emb.n_properties, phys_hidden_channels)
-        else:
-            phys_hidden_channels = self.phys_emb.n_properties
-
-        # Period + group embeddings
-        if self.use_pg:
-            self.period_embedding = Embedding(
-                self.phys_emb.period_size,
-                pg_hidden_channels,
-            )
-            self.group_embedding = Embedding(
-                self.phys_emb.group_size,
-                pg_hidden_channels,
-            )
-
-        # Tag embedding
-        if tag_hidden_channels:
-            self.tag_embedding = Embedding(3, tag_hidden_channels)
-
-        # Main embedding
-        self.emb = Embedding(
-            85,
-            hidden_channels
-            - tag_hidden_channels
-            - phys_hidden_channels
-            - 2 * pg_hidden_channels,
-        )
-
-        # MLP
-        self.lin = Linear(hidden_channels, hidden_channels)
-        if self.second_layer_MLP:
-            self.lin_2 = Linear(hidden_channels, hidden_channels)
-
-        # --- Edge embedding ---
-        self.lin_e1 = Linear(3, num_filters // 2)  # r_ij
-        self.lin_e12 = Linear(num_gaussians, num_filters - (num_filters // 2))  # d_ij
-
-        if self.second_layer_MLP:
-            self.lin_e2 = Linear(num_filters, num_filters)
-
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        self.emb.reset_parameters()
-        if self.use_mlp_phys:
-            nn.init.xavier_uniform_(self.phys_lin.weight)
-        if self.use_tag:
-            self.tag_embedding.reset_parameters()
-        if self.use_pg:
-            self.period_embedding.reset_parameters()
-            self.group_embedding.reset_parameters()
-        nn.init.xavier_uniform_(self.lin.weight)
-        self.lin.bias.data.fill_(0)
-        nn.init.xavier_uniform_(self.lin_e1.weight)
-        self.lin_e1.bias.data.fill_(0)
-        if self.second_layer_MLP:
-            nn.init.xavier_uniform_(self.lin_2.weight)
-            self.lin_2.bias.data.fill_(0)
-            nn.init.xavier_uniform_(self.lin_e2.weight)
-            self.lin_e2.bias.data.fill_(0)
-
-    def forward(self, z, rel_pos, edge_attr, tag=None, subnodes=None):
-        """Forward pass of the Embedding block.
-        Called in FAENet to generate initial atom and edge representations.
-
-        Args:
-            z (tensor): atomic numbers. (num_atoms, )
-            rel_pos (tensor): relative atomic positions. (num_edges, 3)
-            edge_attr (tensor): RBF of pairwise distances. (num_edges, num_gaussians)
-            tag (tensor, optional): atom information specific to OCP. Defaults to None.
-
-        Returns:
-            (tensor, tensor): atom embeddings, edge embeddings
-        """
-
-        # --- Edge embedding --
-        rel_pos = self.lin_e1(rel_pos)  # r_ij
-        edge_attr = self.lin_e12(edge_attr)  # d_ij
-        e = torch.cat((rel_pos, edge_attr), dim=1)
-        e = self.act(e)  # can comment out
-
-        if self.second_layer_MLP:
-            # e = self.lin_e2(e)
-            e = self.act(self.lin_e2(e))
-
-        # --- Node embedding --
-
-        # Create atom embeddings based on its characteristic number
-        h = self.emb(z)
-
-        if self.phys_emb.device != h.device:
-            self.phys_emb = self.phys_emb.to(h.device)
-
-        # Concat tag embedding
-        if self.use_tag:
-            h_tag = self.tag_embedding(tag)
-            h = torch.cat((h, h_tag), dim=1)
-
-        # Concat physics embeddings
-        if self.phys_emb.n_properties > 0:
-            h_phys = self.phys_emb.properties[z]
-            if self.use_mlp_phys:
-                h_phys = self.phys_lin(h_phys)
-            h = torch.cat((h, h_phys), dim=1)
-
-        # Concat period & group embedding
-        if self.use_pg:
-            h_period = self.period_embedding(self.phys_emb.period[z])
-            h_group = self.group_embedding(self.phys_emb.group[z])
-            h = torch.cat((h, h_period, h_group), dim=1)
-
-        # MLP
-        h = self.act(self.lin(h))
-        if self.second_layer_MLP:
-            h = self.act(self.lin_2(h))
-
-        return h, e
-
-
-class InteractionBlock(MessagePassing):
-    """Updates atom representations through custom message passing."""
-
-    def __init__(
-        self,
-        hidden_channels,
-        num_filters,
-        act,
-        mp_type,
-        complex_mp,
-        graph_norm,
-    ):
-        super().__init__()
-        self.act = act
-        self.mp_type = mp_type
-        self.hidden_channels = hidden_channels
-        self.complex_mp = complex_mp
-        self.graph_norm = graph_norm
-        if graph_norm:
-            self.graph_norm = GraphNorm(
-                hidden_channels if "updown" not in self.mp_type else num_filters,
-            )
-
-        if self.mp_type == "simple":
-            self.lin_h = nn.Linear(hidden_channels, hidden_channels)
-
-        elif self.mp_type == "updownscale":
-            self.lin_geom = nn.Linear(num_filters, num_filters)
-            self.lin_down = nn.Linear(hidden_channels, num_filters)
-            self.lin_up = nn.Linear(num_filters, hidden_channels)
-
-        elif self.mp_type == "updownscale_base":
-            self.lin_geom = nn.Linear(num_filters + 2 * hidden_channels, num_filters)
-            self.lin_down = nn.Linear(hidden_channels, num_filters)
-            self.lin_up = nn.Linear(num_filters, hidden_channels)
-
-        elif self.mp_type == "updown_local_env":
-            self.lin_down = nn.Linear(hidden_channels, num_filters)
-            self.lin_geom = nn.Linear(num_filters, num_filters)
-            self.lin_up = nn.Linear(2 * num_filters, hidden_channels)
-
-        else:  # base
-            self.lin_geom = nn.Linear(
-                num_filters + 2 * hidden_channels,
-                hidden_channels,
-            )
-            self.lin_h = nn.Linear(hidden_channels, hidden_channels)
-
-        if self.complex_mp:
-            self.other_mlp = nn.Linear(hidden_channels, hidden_channels)
-
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        if self.mp_type != "simple":
-            nn.init.xavier_uniform_(self.lin_geom.weight)
-            self.lin_geom.bias.data.fill_(0)
-        if self.complex_mp:
-            nn.init.xavier_uniform_(self.other_mlp.weight)
-            self.other_mlp.bias.data.fill_(0)
-        if self.mp_type in {"updownscale", "updownscale_base", "updown_local_env"}:
-            nn.init.xavier_uniform_(self.lin_up.weight)
-            self.lin_up.bias.data.fill_(0)
-            nn.init.xavier_uniform_(self.lin_down.weight)
-            self.lin_down.bias.data.fill_(0)
-        else:
-            nn.init.xavier_uniform_(self.lin_h.weight)
-            self.lin_h.bias.data.fill_(0)
-
-    def forward(self, h, edge_index, e):
-        """Forward pass of the Interaction block.
-        Called in FAENet forward pass to update atom representations.
-
-        Args:
-            h (tensor): atom embedddings. (num_atoms, hidden_channels)
-            edge_index (tensor): adjacency matrix. (2, num_edges)
-            e (tensor): edge embeddings. (num_edges, num_filters)
-
-        Returns:
-            (tensor): updated atom embeddings
-        """
-        # Define edge embedding
-        if self.mp_type in {"base", "updownscale_base"}:
-            e = torch.cat([e, h[edge_index[0]], h[edge_index[1]]], dim=1)
-
-        if self.mp_type in {
-            "updownscale",
-            "base",
-            "updownscale_base",
-        }:
-            e = self.act(self.lin_geom(e))
-
-        # --- Message Passing block --
-
-        if self.mp_type == "updownscale" or self.mp_type == "updownscale_base":
-            h = self.act(self.lin_down(h))  # downscale node rep.
-            h = self.propagate(edge_index, x=h, W=e)  # propagate
-            if self.graph_norm:
-                h = self.act(self.graph_norm(h))
-            h = self.act(self.lin_up(h))  # upscale node rep.
-
-        elif self.mp_type == "updown_local_env":
-            h = self.act(self.lin_down(h))
-            chi = self.propagate(edge_index, x=h, W=e, local_env=True)
-            e = self.lin_geom(e)
-            h = self.propagate(edge_index, x=h, W=e)  # propagate
-            if self.graph_norm:
-                h = self.act(self.graph_norm(h))
-            h = torch.cat((h, chi), dim=1)
-            h = self.lin_up(h)
-
-        elif self.mp_type in {"base", "simple"}:
-            h = self.propagate(edge_index, x=h, W=e)  # propagate
-            if self.graph_norm:
-                h = self.act(self.graph_norm(h))
-            h = self.act(self.lin_h(h))
-
-        else:
-            raise ValueError("mp_type provided does not exist")
-
-        if self.complex_mp:
-            h = self.act(self.other_mlp(h))
-
-        return h
-
-    def message(self, x_j, W, local_env=None):
-        if local_env is not None:
-            return W
-        else:
-            return x_j * W
-
-
-class OutputBlock(nn.Module):
-    """Compute task-specific predictions from final atom representations."""
-
-    def __init__(self, energy_head, hidden_channels, act, out_dim=1):
-        super().__init__()
-        self.energy_head = energy_head
-        self.act = act
-
-        self.lin1 = Linear(hidden_channels, hidden_channels // 2)
-        self.lin2 = Linear(hidden_channels // 2, out_dim)
-
-        if self.energy_head == "weighted-av-final-embeds":
-            self.w_lin = Linear(hidden_channels, 1)
-
-    def reset_parameters(self):
-        nn.init.xavier_uniform_(self.lin1.weight)
-        self.lin1.bias.data.fill_(0)
-        nn.init.xavier_uniform_(self.lin2.weight)
-        self.lin2.bias.data.fill_(0)
-        if self.energy_head == "weighted-av-final-embeds":
-            nn.init.xavier_uniform_(self.w_lin.weight)
-            self.w_lin.bias.data.fill_(0)
-
-    def forward(self, h, edge_index, edge_weight, batch, alpha):
-        """Forward pass of the Output block.
-        Called in FAENet to make prediction from final atom representations.
-
-        Args:
-            h (tensor): atom representations. (num_atoms, hidden_channels)
-            edge_index (tensor): adjacency matrix. (2, num_edges)
-            edge_weight (tensor): edge weights. (num_edges, )
-            batch (tensor): batch indices. (num_atoms, )
-            alpha (tensor): atom attention weights for late energy head. (num_atoms, )
-
-        Returns:
-            (tensor): graph-level representation (e.g. energy prediction)
-        """
-        if self.energy_head == "weighted-av-final-embeds":
-            alpha = self.w_lin(h)
-
-        # MLP
-        h = self.lin1(h)
-        h = self.act(h)
-        h = self.lin2(h)
-
-        if self.energy_head in {
-            "weighted-av-initial-embeds",
-            "weighted-av-final-embeds",
-        }:
-            h = h * alpha
-
-        # Global pooling
-        out = torch.scatter(h, batch, dim=0, reduce="add")
-
-        return out
 
 
 class FAENet(AbstractPyGModel):
@@ -560,8 +216,37 @@ class FAENet(AbstractPyGModel):
                 self.hidden_channels,
             )
 
-    # FAENet's forward pass in done in BaseModel, inherited here.
-    # It uses forces_forward() and energy_forward() defined below.
+    def read_batch(self, batch: BatchDict) -> DataDict:
+        r"""
+        Extracts data needed by FAENet from the batch and graph
+        structures.
+
+        Node features
+        Edge features
+
+        Parameters
+        ----------
+        batch : BatchDict
+            Batch of data to be processed
+
+        Returns
+        -------
+        DataDict
+            Input data for FAENet as a dictionary.
+        """
+
+        data = super().read_batch(batch)
+        data["graph"].cell = batch["cell"]
+        data["graph"].natoms = batch["natoms"].squeeze(-1).to(torch.int32)
+        edge_index, cell_offsets, neighbors = radius_graph_pbc(
+            data["graph"],
+            self.cutoff,
+            50,
+        )
+        data["graph"].edge_index = edge_index
+        data["graph"].cell_offsets = cell_offsets
+        data["graph"].neighbors = neighbors
+        return data
 
     def forces_forward(self, preds):
         """Predicts forces for 3D atomic systems.
@@ -593,9 +278,7 @@ class FAENet(AbstractPyGModel):
         # Pre-process data (e.g. pbc, cutoff graph, etc.)
         # Should output all necessary attributes, in correct format.
         if preproc:
-            import pdb
-
-            pdb.set_trace()
+            # import pdb; pdb.set_trace()
             z, batch, edge_index, rel_pos, edge_weight = self.preprocess(
                 data,
                 self.cutoff,
@@ -657,7 +340,15 @@ class FAENet(AbstractPyGModel):
 
         return preds
 
-    def forward(self, data, mode="train", preproc=True):
+    # def _forward(self, data, mode="train", preproc=True):
+    def _forward(
+        self,
+        graph: AbstractGraph,
+        node_feats: torch.Tensor,
+        pos: torch.Tensor,
+        edge_feats: torch.Tensor | None = None,
+        **kwargs,
+    ) -> torch.Tensor:
         """Main Forward pass.
 
         Args:
@@ -669,7 +360,9 @@ class FAENet(AbstractPyGModel):
         Returns:
             (dict): predicted energy, forces and final atomic hidden states
         """
-        data = data["graph"]
+        mode = "train"
+        preproc = True
+        data = graph
         grad_forces = forces = None
 
         # energy gradient w.r.t. positions will be computed
@@ -712,22 +405,22 @@ class FAENet(AbstractPyGModel):
 
         return preds
 
-    # def _forward(
-    #     self,
-    #     batch,
-    #     model,
-    #     frame_averaging,
-    #     mode="train",
-    #     crystal_task=True,
-    # ):
-    def _forward(
+    def final_forward(
         self,
-        graph: AbstractGraph,
-        node_feats: torch.Tensor,
-        pos: torch.Tensor,
-        edge_feats: torch.Tensor | None = None,
-        **kwargs,
-    ) -> torch.Tensor:
+        batch,
+        model,
+        frame_averaging,
+        mode="train",
+        crystal_task=True,
+    ):
+        # def _forward(
+        #     self,
+        #     graph: AbstractGraph,
+        #     node_feats: torch.Tensor,
+        #     pos: torch.Tensor,
+        #     edge_feats: torch.Tensor | None = None,
+        #     **kwargs,
+        # ) -> torch.Tensor:
         """Perform a model forward pass when frame averaging is applied.
 
         Args:
@@ -751,9 +444,6 @@ class FAENet(AbstractPyGModel):
         frame_averaging = "3D"
         mode = "train"
         crystal_task = True
-        import pdb
-
-        pdb.set_trace()
 
         if isinstance(batch, list):
             batch = batch[0]
